Move filteringProwJobLister into JobAgent
There is no reason to ever use the JobAgent without the filtering
ProwJobLister and its the only thing that uses it.

This change is needed so external consumers can use the JobAgent.
