+++ prow/deck/jobs/jobs.go
@@ -19,6 +19,7 @@ package jobs
 
 import (
 	"bytes"
+	"context"
 	"errors"
 	"fmt"
 	"io/ioutil"
@@ -30,6 +31,9 @@ import (
 
 	"github.com/sirupsen/logrus"
 	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/util/sets"
+	ctrlruntimeclient "sigs.k8s.io/controller-runtime/pkg/client"
+
 	prowapi "k8s.io/test-infra/prow/apis/prowjobs/v1"
 	"k8s.io/test-infra/prow/config"
 )
@@ -78,13 +82,74 @@ type PodLogClient interface {
 	GetLogs(name string) ([]byte, error)
 }
 
+// PJListingClient is an interface to list ProwJobs
+type PJListingClient interface {
+	List(context.Context, *prowapi.ProwJobList, ...ctrlruntimeclient.ListOption) error
+}
+
 // NewJobAgent is a JobAgent constructor.
-func NewJobAgent(kc serviceClusterClient, plClients map[string]PodLogClient, cfg config.Getter) *JobAgent {
+func NewJobAgent(ctx context.Context, pjLister PJListingClient, hiddenOnly, showHidden bool, plClients map[string]PodLogClient, cfg config.Getter) *JobAgent {
 	return &JobAgent{
-		kc:     kc,
+		kc: &filteringProwJobLister{
+			ctx:         ctx,
+			client:      pjLister,
+			hiddenRepos: func() sets.String { return sets.NewString(cfg().Deck.HiddenRepos...) },
+			hiddenOnly:  hiddenOnly,
+			showHidden:  showHidden,
+			cfg:         cfg,
+		},
 		pkcs:   plClients,
 		config: cfg,
 	}
+}
+
+type filteringProwJobLister struct {
+	ctx         context.Context
+	client      PJListingClient
+	cfg         config.Getter
+	hiddenRepos func() sets.String
+	hiddenOnly  bool
+	showHidden  bool
+}
+
+func (c *filteringProwJobLister) ListProwJobs(selector string) ([]prowapi.ProwJob, error) {
+	prowJobList := &prowapi.ProwJobList{}
+	parsedSelector, err := labels.Parse(selector)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse selector: %v", err)
+	}
+	listOpts := &ctrlruntimeclient.ListOptions{LabelSelector: parsedSelector, Namespace: c.cfg().ProwJobNamespace}
+	if err := c.client.List(c.ctx, prowJobList, listOpts); err != nil {
+		return nil, err
+	}
+
+	var filtered []prowapi.ProwJob
+	for _, item := range prowJobList.Items {
+		shouldHide := item.Spec.Hidden || c.pjHasHiddenRefs(item)
+		if shouldHide && c.showHidden {
+			filtered = append(filtered, item)
+		} else if shouldHide == c.hiddenOnly {
+			// this is a hidden job, show it if we're asked
+			// to only show hidden jobs otherwise hide it
+			filtered = append(filtered, item)
+		}
+	}
+
+	return filtered, nil
+}
+
+func (c *filteringProwJobLister) pjHasHiddenRefs(pj prowapi.ProwJob) bool {
+	allRefs := pj.Spec.ExtraRefs
+	if pj.Spec.Refs != nil {
+		allRefs = append(allRefs, *pj.Spec.Refs)
+	}
+	for _, refs := range allRefs {
+		if c.hiddenRepos().HasAny(fmt.Sprintf("%s/%s", refs.Org, refs.Repo), refs.Org) {
+			return true
+		}
+	}
+
+	return false
 }
 
 // JobAgent creates lists of jobs, updates their status and returns their run logs.
