+++ prow/spyglass/BUILD.bazel
@@ -26,6 +26,7 @@ go_test(
         "@com_github_googlecloudplatform_testgrid//pb/config:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
+        "@io_k8s_sigs_controller_runtime//pkg/client:go_default_library",
     ],
 )
